jquery cdn added, DeezerAPI calls separated
