+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "mp3-metadata-app",
-  "version": "0.3.0",
+  "version": "0.3.1",
   "private": true,
   "author": {
     "name": "Benjamin Fajic",
