+++ public/index.html
@@ -30,6 +30,7 @@
     <div id="app"></div>
     <!-- built files will be auto injected -->
     <div id="dz-root"></div>
+    <script src="http://ajax.googleapis.com/ajax/libs/jquery/1.7.1/jquery.min.js" type="text/javascript"></script>
     <script src="https://e-cdns-files.dzcdn.net/js/min/dz.js"></script>
   </body>
 </html>
